Add thin wrappers around C API bindings
These do not provide much in the way of type-safety so far, but are at
least more pleasant to work with than the raw C API.  Haddocks for
them are copy-pasted from TDLib docs, with minimal modifications.

Also, add a simple test program using some of them, just to verify
that the library works in principle.