Beginnings of a more reasonable API
The use of TemplateHaskell for deriving aeson instances is motivated
primarily by DRY considerations, not efficiency.  Specifically, since
we have to use custom options to match tdlib's scheme, it is way
easier to pass these options once into deriveJSON than duplicate them
for each generically-derived method.  The fact that this way also
produces faster code is merely an icing on the cake.