Add fieldLabelModifier
Currently, there are two parts to it: converting camelCase to
snake_case, because the former is the usual convention in Haskell, and
the latter is used by TDLib API; and second, drop leading
underscores.  The second part is a bit of a violation of YAGNI, but is
added in anticipation of future lens shenanigans.